+++ R/prediction.R
@@ -2,11 +2,11 @@
 #'
 #' @param data data to predict (`data.frame`/`matrix` format), from
 #' [gcap.collapse2Genes()] in general.
+#' @param model model name ("XGB11", "XGB32", "XGB54",
+#' "XGB11_stepwise", "XGB32_stepwise", "XGB54_stepwise") or a custom model
+#' from input. 'toy' can be used for test.
 #' @param target one of 'circle' and 'nonLinear' to select model
 #' if predict circle amplicon or non-linear amplicon.
-#' @param use_toy if `TRUE`, use a toy model for showing example,
-#' not for practical use.
-#' @param custom_model a outer custom xgboost model for prediction.
 #'
 #' @return a numeric vector representing prob.
 #' @importFrom stats predict
@@ -14,48 +14,43 @@
 #'
 #' @examples
 #' data("ec")
-#' y_pred <- gcap.runPrediction(ec)
+#' # Use toy model for illustration
+#' y_pred <- gcap.runPrediction(ec, "toy")
 #' y_pred
 #' @testexamples
 #' expect_equal(length(y_pred), 2020L)
 gcap.runPrediction <- function(data,
-                               target = c("circle", "nonLinear"),
-                               use_toy = FALSE,
-                               custom_model = NULL) {
+                               model = "XGB32",
+                               target = c("circle", "nonLinear")) {
   stopifnot(data.table::is.data.table(data))
   target <- match.arg(target)
 
   lg <- set_logger()
 
   lg$info("reading trained model based on input target setting: {target}")
-  if (use_toy) {
-    model <- readRDS(
-      system.file(
-        "extdata", "toy_model.rds",
-        package = "gcap", mustWork = TRUE
+  if (is.character(model)) {
+    if (model == "toy") {
+      model <- readRDS(
+        system.file(
+          "extdata", "toy_model.rds",
+          package = "gcap", mustWork = TRUE
+        )
       )
-    )
-  } else {
-    if (!is.null(custom_model)) {
-      lg$info("a custom model is selected from user input")
-      model <- custom_model
     } else {
-      if (all(c("age", "gender") %in% colnames(data))) {
-        fl <- c(
-          "final_xgb_model_circle_without_type.rds",
-          "final_xgb_model_nonLinear_without_type.rds"
+      if (target == "circle") {
+        modfile <- switch(model,
+          XGB11 = "xgb_stepwise_model_NF11.rds", # NEED UPDATE
+          XGB32 = "xgb_stepwise_model_NF11.rds",
+          XGB54 = "xgb_stepwise_model_NF11.rds",
+          XGB11_stepwise = "xgb_stepwise_model_NF11.rds",
+          XGB32_stepwise = "xgb_stepwise_model_NF32.rds",
+          XGB54_stepwise = "xgb_stepwise_model_NF54.rds",
+          stop("Unsupported model input!")
         )
       } else {
-        lg$info("without 'age' and 'gender' detected, an alternative model will be used")
-        fl <- c(
-          "final_xgb_model_circle_without_type_rm_cli.rds",
-          "final_xgb_model_nonLinear_without_type_rm_cli.rds"
-        )
+        stop("Currently not supported!")
       }
 
-      names(fl) <- c("circle", "nonLinear")
-
-      modfile <- as.character(fl[target])
       lg$info("using model file {modfile}")
 
       model <- readRDS(
@@ -65,6 +60,8 @@ gcap.runPrediction <- function(data,
         )
       )
     }
+  } else {
+    lg$info("a custom model is selected from user input")
   }
 
   lg$info("selecting necessary features from input data")